visual improvements to TasksGrid and move TasksGridColumn component to it's own file. tasksGridColumns.js file to specify which columns should be rendered and corresponding enum for graphql
