+++ src/scenes/Dashboard/components/tasksGridColumns.js
@@ -1,0 +1,10 @@
+const tasksGridColumns = {
+    tasksNew: "NEW",
+    tasksActive: "ACTIVE",
+    tasksPickedUp: "PICKED_UP",
+    tasksDelivered: "DROPPED_OFF",
+    tasksCancelled: "CANCELLED",
+    tasksRejected: "REJECTED",
+};
+
+export default tasksGridColumns;
